fixed gTimer bug

+++ Timer.cpp
@@ -16,7 +16,7 @@ float gTimer::getDelta()
 	LARGE_INTEGER current;
 	QueryPerformanceCounter( &current );
 
-	float delta = ( (float)current.QuadPart - (float)m_last.QuadPart ) / (float)m_freq.QuadPart;
+	float delta = ( current.QuadPart - m_last.QuadPart ) / (float)m_freq.QuadPart;
 	m_last = current;
 	return delta;
 }
